feat: change the default image
Rather than a 16x16 pixels image,
use one with 3x2 pixels that has the aspect ratio used at der Freitag.

This should avoid the reflowings on page listings. As the rectangular
image (placeholder) was loaded first, layout was set, and afterwards the
final image was loaded with a different aspect ratio, making the layout
not reflow properly (a bug on Foundation CSS by not using the latest
grid).